+++ argocd/outputs.tf
@@ -1,0 +1,4 @@
+output "argocd-metadata" {
+  description = "The details of the manifests used to deploy ArgoCD."
+  value       = helm_release.argocd
+}